+++ clinica/iotools/utils/data_handling.py
@@ -505,6 +505,14 @@ def compute_missing_processing(bids_dir, caps_dir, out_file):
                 else:
                     row_df.loc[0, f"pet-surface_{trc}"] = "0"
 
+            # Check pet-linear outputs
+            for trc in trc_avail:
+                pet_pattern = path.join(session_path, "pet_linear", f"*{trc}*")
+                if len(glob(pet_pattern)) > 0:
+                    row_df.loc[0, f"pet-linear_{trc}"] = "1"
+                else:
+                    row_df.loc[0, f"pet-linear_{trc}"] = "0"
+
             output_df = pd.concat([output_df, row_df])
 
     output_df.sort_values(["participant_id", "session_id"], inplace=True)
@@ -714,6 +722,7 @@ def create_subs_sess_list(
             session_df = pd.read_csv(
                 path.join(sub_path, subj_id + "_sessions.tsv"), sep="\t"
             )
+            session_df.dropna(how="all", inplace=True)
             session_list = list(session_df["session_id"].to_numpy())
             for session in session_list:
                 subjs_sess_tsv.write(subj_id + "\t" + session + "\n")