+++ test/nonregression/iotools/test_run_converters.py
@@ -7,17 +7,9 @@ different functions available in Clinica
 
 
 import warnings
-from os import PathLike, fspath
+from os import PathLike
 from pathlib import Path
-from test.nonregression.testing_tools import (
-    clean_folder,
-    compare_folders,
-    compare_folders_structures,
-    compare_folders_with_hashes,
-    create_list_hashes,
-    identical_subject_list,
-    same_missing_modality_tsv,
-)
+from test.nonregression.testing_tools import compare_folders
 
 import pytest
 
@@ -27,6 +19,7 @@ warnings.filterwarnings("ignore")
 
 @pytest.fixture(
     params=[
+        # TODO: Update NIFD reference dataset.
         "Nifd2Bids",
         "Oasis2Bids",
         "Oasis3ToBids",
@@ -41,32 +34,39 @@ def test_name(request):
 
 
 def run_nifd2bids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) -> None:
-    import shutil
+    from pathlib import PurePath
+    from tempfile import TemporaryDirectory
 
     from clinica.iotools.converters.nifd_to_bids.nifd_to_bids import convert_images
 
-    # Arrange
-    shutil.copytree(
-        input_dir / "clinical_data",
-        output_dir / "clinical_data",
-        copy_function=shutil.copy,
+    # Convert
+    input_dir = PurePath(input_dir)
+    output_dir = PurePath(output_dir)
+    ref_dir = PurePath(ref_dir)
+
+    # Act
+    _ = convert_images(
+        path_to_clinical=input_dir / "clinical_data",
+        path_to_dataset=input_dir / "unorganized",
+        bids_dir=output_dir,
     )
-    # Arrange - Data location
-    clinical_data_directory = output_dir / "clinical_data"
-    # Acte - Conversion
-    to_convert = convert_images(
-        input_dir / "unorganized", output_dir / "bids", clinical_data_directory
-    )
+
     # Assert
-    compare_folders_structures(
-        fspath(output_dir / "bids"), fspath(ref_dir / "hashes_nifd.p")
-    )
+    with TemporaryDirectory() as td:
+        compare_folders(output_dir, ref_dir, td)
 
 
 def run_oasis2bids(
     input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike
 ) -> None:
+    from pathlib import PurePath
+
     from clinica.iotools.converters.oasis_to_bids.oasis_to_bids import OasisToBids
+
+    # Convert
+    input_dir = PurePath(input_dir)
+    output_dir = PurePath(output_dir)
+    ref_dir = PurePath(ref_dir)
 
     # Arrange
     clinical_data_directory = input_dir / "clinical_data"
@@ -81,20 +81,36 @@ def run_oasis2bids(
 def run_oasis3tobids(
     input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike
 ) -> None:
+    from pathlib import PurePath
+
     from clinica.iotools.converters.oasis3_to_bids.oasis3_to_bids import convert_images
+
+    # Convert
+    input_dir = PurePath(input_dir)
+    output_dir = PurePath(output_dir)
+    ref_dir = PurePath(ref_dir)
 
     # Arrange
     clinical_data_directory = input_dir / "clinical_data"
+
     # Act
     convert_images(
         input_dir / "unorganized", output_dir / "bids", clinical_data_directory
     )
+
     # Assert
     compare_folders(output_dir / "bids", ref_dir / "bids", output_dir)
 
 
 def run_adni2bids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) -> None:
+    from pathlib import PurePath
+
     from clinica.iotools.converters.adni_to_bids.adni_to_bids import AdniToBids
+
+    # Convert
+    input_dir = PurePath(input_dir)
+    output_dir = PurePath(output_dir)
+    ref_dir = PurePath(ref_dir)
 
     # Arrange
     clinical_data_directory = input_dir / "clinical_data"
@@ -102,6 +118,7 @@ def run_adni2bids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) 
     dataset_directory = input_dir / "unorganized_data"
     subjects_list = input_dir / "subjects.txt"
     modalities = ["T1", "PET_FDG", "PET_AMYLOID", "PET_TAU", "DWI", "FLAIR", "fMRI"]
+
     # Act
     adni_to_bids = AdniToBids()
     adni_to_bids.check_adni_dependencies()
@@ -117,19 +134,28 @@ def run_adni2bids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) 
         output_dir / "bids",
         xml_path=xml_directory,
     )
+
     # Assert
     compare_folders(output_dir / "bids", ref_dir / "bids", output_dir)
 
 
 def run_aibl2bids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) -> None:
+    from pathlib import PurePath
+
     from clinica.iotools.converters.aibl_to_bids.aibl_to_bids import (
         convert_clinical_data,
         convert_images,
     )
 
+    # Convert
+    input_dir = PurePath(input_dir)
+    output_dir = PurePath(output_dir)
+    ref_dir = PurePath(ref_dir)
+
     # Arrange
     clinical_data_directory = input_dir / "Data_extract_3.2.5"
     dataset_directory = input_dir / "unorganized_data"
+
     # Act
     convert_images(
         dataset_directory,
@@ -137,11 +163,14 @@ def run_aibl2bids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) 
         output_dir / "bids",
     )
     convert_clinical_data(output_dir / "bids", clinical_data_directory)
+
     # Assert
     compare_folders(output_dir / "bids", ref_dir / "bids", output_dir)
 
 
-def run_habs_to_bids(input_dir: Path, output_dir: Path, ref_dir: Path):
+def run_habs_to_bids(
+    input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike
+) -> None:
     from click.testing import CliRunner
 
     from clinica.iotools.converters.habs_to_bids.habs_to_bids_cli import cli
@@ -154,23 +183,30 @@ def run_habs_to_bids(input_dir: Path, output_dir: Path, ref_dir: Path):
 
 
 def run_ukbtobids(input_dir: PathLike, output_dir: PathLike, ref_dir: PathLike) -> None:
+    from pathlib import PurePath
+
     from clinica.iotools.converters.ukb_to_bids.ukb_to_bids import convert_images
     from clinica.utils.check_dependency import check_dcm2niix
 
+    # Convert
+    input_dir = PurePath(input_dir)
+    output_dir = PurePath(output_dir)
+    ref_dir = PurePath(ref_dir)
+
     # Arrange
     clinical_data_directory = input_dir / "clinical_data"
+
     # Act
     check_dcm2niix()
     convert_images(
         input_dir / "unorganized", output_dir / "bids", clinical_data_directory
     )
+
     # Assert
     compare_folders(output_dir / "bids", ref_dir / "bids", output_dir / "bids")
 
 
 def test_run_convertors(cmdopt, tmp_path, test_name):
-    import shutil
-
     base_dir = Path(cmdopt["input"])
     input_dir = base_dir / test_name / "in"
     ref_dir = base_dir / test_name / "ref"